refactor(hip): trim shape foundation scaffolding
Keep the foundation independently exercised by moving deferred contracts to their first consumers and isolating test-only passes from production tools.

Made-with: Cursor
