+++ lib/Dialect/IR/CMakeLists.txt
@@ -10,9 +10,6 @@ add_library(HipDialectIR STATIC
   HipOpStateInterface.cpp
   HipReifyResultShapesImpl.cpp
   HipShapeUtils.cpp
-  HipShapeUtilsInternal.cpp
-  HipShapeUtilsAttention.cpp
-  HipShapeUtilsConvPool.cpp
   HipShapeUtilsGather.cpp
   HipShapeUtilsMatmulGemm.cpp
   HipShapeUtilsReduction.cpp
@@ -47,10 +44,3 @@ target_link_libraries(HipDialectIR PUBLIC
   # (the per-op state constructor calls woven into the op-state init function).
   MLIRLLVMDialect
 )
-
-# HipDialect.cpp aggregates every TableGen-generated op class (GET_OP_CLASSES).
-# As the dialect grows, this single translation unit exceeds the MSVC COFF
-# section limit (fatal error C1128); /bigobj raises it.
-if(MSVC)
-  target_compile_options(HipDialectIR PRIVATE /bigobj)
-endif()
