--- lib/Dialect/IR/HipShapeUtilsAttention.cpp
@@ -1,7 +1,0 @@
-/*
- * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
- * Licensed under the MIT License.
- */
-//===- HipShapeUtilsAttention.cpp - Attention shape helpers --------------===//
-
-#include "hip/Dialect/IR/HipShapeUtils.h"
