--- lib/Dialect/IR/HipShapeUtilsConvPool.cpp
@@ -1,7 +1,0 @@
-/*
- * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
- * Licensed under the MIT License.
- */
-//===- HipShapeUtilsConvPool.cpp - Conv/pool shape helpers ---------------===//
-
-#include "hip/Dialect/IR/HipShapeUtils.h"
